% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_functions.R
\name{adjust_plotly_facet_axis_titles}
\alias{adjust_plotly_facet_axis_titles}
\title{Adjust Location of Plotly's Facet X & Y Axis Titles}
\usage{
adjust_plotly_facet_axis_titles(p, x = NULL, y = NULL)
}
\arguments{
\item{p}{A plotly object.}

\item{x}{A numeric. Value to place x-axis title on a plotly object. Use \code{get_plotly_axis_titles_values()} to get the objects current value as a reference.}

\item{y}{A numeric. Value to place y-axis title on a plotly object. Use \code{get_plotly_axis_titles_values()} to get the objects current value as a reference.}
}
\description{
Credit: https://github.com/ropensci/plotly/issues/1224
}
\examples{
\dontrun{
p \%>\%
  adjust_plotly_axis_titles(x = -0.15)
}
}
