% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_when_factor.R
\name{case_when_factor}
\alias{case_when_factor}
\title{Turn case_when output into an ordered factor}
\usage{
case_when_factor(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> A sequence of two-sided formulas. The left hand side (LHS)
determines which values match this case. The right hand side (RHS)
provides the replacement value.

The LHS must evaluate to a logical vector. The RHS does not need to be
logical, but all RHSs must evaluate to the same type of vector.

Both LHS and RHS may have the same length of either 1 or \code{n}. The
value of \code{n} must be consistent across all cases. The case of
\code{n == 0} is treated as a variant of \code{n != 1}.

\code{NULL} inputs are ignored.}
}
\value{
An ordered factor vector of length 1 or n, matching the length of the logical input or output vectors, with the type (and attributes) of the first RHS. Inconsistent lengths or types will generate an error.
}
\description{
Turn case_when output into an ordered factor
}
\examples{
library(dplyr)
mtcars2 <-
  mtcars \%>\%
  mutate(
    mpg_bin = case_when(
      mpg < 10 ~ "mpg < 10",
      mpg >= 10 & mpg < 20 ~ "10 <= mpg < 20",
      TRUE ~ "20 <= mpg"
    ),
    mpg_bin_factor = case_when_factor(
      mpg < 10 ~ "mpg < 10",
      mpg >= 10 & mpg < 20 ~ "10 <= mpg < 20",
      TRUE ~ "20 <= mpg"
    )
  ) \%>\%
  select(mpg, starts_with("mpg_bin")) \%>\%
  as_tibble()
mtcars2
class(mtcars2$mpg_bin)
class(mtcars2$mpg_bin_factor)
levels(mtcars2$mpg_bin_factor)
}
\references{
https://stackoverflow.com/questions/49572416/r-convert-to-factor-with-order-of-levels-same-with-case-when
}
