% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ggplot_axis_numbers.R
\name{clean_ggplot_axis_numbers}
\alias{clean_ggplot_axis_numbers}
\title{Scale Down Numbers for Plots}
\usage{
clean_ggplot_axis_numbers(x, prefix = NULL)
}
\arguments{
\item{x}{A numeric.}

\item{prefix}{A character. This will be the leading symbol of the final value. Example being a $ symbol.}
}
\description{
This code is a modified from this source code: https://stackoverflow.com/questions/28159936/formatting-large-currency-or-dollar-values-to-millions-billions
}
