% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_functions.R
\name{get_plotly_facet_axis_titles_values}
\alias{get_plotly_facet_axis_titles_values}
\title{Get Plotly's Facet X & Y Axis Title Values}
\usage{
get_plotly_facet_axis_titles_values(p)
}
\arguments{
\item{p}{A plotly object.}
}
\description{
Credit: https://github.com/ropensci/plotly/issues/1224
}
\examples{
\dontrun{
p \%>\%
  get_plotly_axis_title_values()
}
}
