% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_kow.R
\name{kow_pallette}
\alias{kow_pallette}
\title{Kow color palette (discrete)}
\usage{
kow_pallette()
}
\description{
Color palettes for me.
This palette includes 7 colors.
}
\seealso{
Other colour kow: \code{\link{scale_color_kow}}
}
\concept{colour kow}
