% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kowr.R
\docType{package}
\name{kowr-package}
\alias{kowr}
\alias{kowr-package}
\title{kowr: Utility Functions}
\description{
This is the realm where the random chunks of code I use will
    live.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://koderkow.github.io/kowr}
  \item \url{https://github.com/KoderKow/kowr}
  \item Report bugs at \url{https://github.com/KoderKow/kowr/issues}
}

}
\author{
\strong{Maintainer}: Kyle Harris \email{koderkow@gmail.com}

}
\keyword{internal}
