% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_group_split.R
\name{named_group_split}
\alias{named_group_split}
\title{Named Split data frame by group}
\usage{
named_group_split(.tbl, ..., arrange = NULL)
}
\arguments{
\item{.tbl}{A tbl}

\item{...}{Grouping specification, forwarded to \code{\link[dplyr:group_by]{dplyr::group_by()}}}

\item{arrange}{A character. If \code{NULL}, the default, this will sort the list alphabetically. Use \code{asc} or \code{desc}if sorting the list by count is needed.}
}
\value{
returns a named list of tibbles. Each tibble contains the rows of .tbl for the associated group and all the columns, including the grouping variables.
}
\description{
Named Split data frame by group
}
\references{
https://github.com/tidyverse/dplyr/issues/4223
}
