% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Not In Operator}
\usage{
lhs \%not_in\% rhs
}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\description{
Opposite of \code{\%in\%} to avoid the \strong{!} in \code{filter(!column \%in\% element)}.
}
