% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{not_na}
\alias{not_na}
\title{Detect if Not NA}
\usage{
not_na(x)
}
\arguments{
\item{x}{an R object to be tested: the default method for is.na and anyNA handle atomic vectors, lists, pairlists, and NULL.}
}
\description{
Opposite of \code{is.na(...)}. Avoiding \strong{!} for cleaner code.
}
