% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%replace_null\%}
\alias{\%replace_null\%}
\title{Replace Null With ...}
\usage{
lhs \%replace_null\% rhs
}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\description{
A quick method to quickly handle \code{NULL} values. Thanks Colin!
}
