% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducability_receipt.R
\name{reproducability_receipt}
\alias{reproducability_receipt}
\title{Add a reproducibility receipt to Rmarkdown}
\usage{
reproducability_receipt()
}
\value{
Invisible NULL; code copied to the users clipboard.
}
\description{
Adds code to your clipboard that you can use to add a "reproducibility receipt" to your Rmarkdown documents.
}
\examples{
\dontrun{
## Add code to your clipboard to copy and paste
reproducibility_receipt()
}
}
\references{
https://twitter.com/MilesMcBain/status/1263272935197782016?s=20
}
