% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_preserve_sum.R
\name{round_preserve_sum}
\alias{round_preserve_sum}
\title{Round + Preserve the Sum}
\usage{
round_preserve_sum(x, digits = 4)
}
\arguments{
\item{x}{A numeric vector to apply the algorithm to. This should be a proportion vector where all values equal 1.}

\item{digits}{A numeric. Indicates the number of decimal places to be used.}
}
\description{
\enumerate{
\item Round down to the specified number of decimal places
\item Order numbers by their remainder values
\item Increment the specified decimal place of values with ‘k’ largest remainders, where ‘k’ is the number of values that must be incremented to preserve their rounded sum
}
}
\examples{
sum(c(0.333, 0.333, 0.334))

sum(round(c(0.333, 0.333, 0.334), 2))

sum(round_preserve_sum(c(0.333, 0.333, 0.334), 2))
}
\references{
https://www.r-bloggers.com/2016/07/round-values-while-preserve-their-rounded-sum-in-r/
}
