% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_up.R
\name{round_up}
\alias{round_up}
\title{Round Up}
\usage{
round_up(x, format = "normal")
}
\arguments{
\item{x}{A numeric value.}

\item{format}{A boolean. Takes three different options to round up and stylize:
\itemize{
\item normal: The default, return the number with no formatting.
\item dollar: Return the number with a leading dollar sign along with commas.
\item comma: Return the number with commas.
}}
}
\value{
A numeric, or a string if format is not "normal".
}
\description{
Round a number up to the number place second to the left. This is useful for reports that need rounded minimums and maximums.
}
\examples{
\dontrun{
# Round a number up
round_up(123456)

# Number with only 2 numeric places:
round_up(19, format = "dollar")
}
}
