% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kowr-data.R
\docType{data}
\name{sample_dates}
\alias{sample_dates}
\title{All dates from the year 2000}
\format{
A data frame with 366 rows and 3 variables:
\describe{
\item{date}{date from the year 2000}
\item{x}{random value 0-100}
\item{y}{random value 0-100}
}
}
\source{
\url{https://github.com/KoderKow/kowr/blob/master/data-raw/zodiac.R}
}
\usage{
sample_dates
}
\description{
This dataset consists of all dates from the year 2000. There are two additional columns to help build sample plots.
}
\keyword{datasets}
