% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zodiac_sign.R
\name{scale_color_zodiac}
\alias{scale_color_zodiac}
\title{Discrete zodiac colors}
\usage{
scale_color_zodiac(..., use_factor_order = FALSE)
}
\arguments{
\item{...}{Extra parameters to pass to \code{ggplot2::discrete_scale()} when \code{use_factor_order} is \code{FALSE} (default).}

\item{use_factor_order}{A logical. If \code{TRUE}, this will assign colors based on how they match with the zodiac sign. If \code{FALSE} (default) then the colors will be used without looking for zodiac sign values.}
}
\description{
Discrete zodiac colors
}
\examples{
library(dplyr)
library(ggplot2)

## Use colors that match with each zodiac sign
sample_dates \%>\%
  mutate(
    zodiac_sign = zodiac_sign(date)
  ) \%>\%
  ggplot() +
  aes(
    x = x,
    y = y,
    color = zodiac_sign
  ) +
  geom_point() +
  ## Set use_factor_order = TRUE
  scale_color_zodiac(use_factor_order = TRUE)

## Use the colors in the palette without zodiac signs
iris \%>\%
  ggplot() +
  aes(
    x = Sepal.Length,
    y = Sepal.Width,
    color = Species
  ) +
  geom_point() +
  scale_color_zodiac()
}
