% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zodiac_sign.R
\name{scale_fill_zodiac}
\alias{scale_fill_zodiac}
\title{Discrete zodiac colors}
\usage{
scale_fill_zodiac(..., use_factor_order = FALSE)
}
\arguments{
\item{...}{Extra parameters to pass to \code{ggplot2::discrete_scale()} when \code{use_factor_order} is \code{FALSE} (default).}

\item{use_factor_order}{A logical. If \code{TRUE}, this will assign colors based on how they match with the zodiac sign. If \code{FALSE} (default) then the colors will be used without looking for zodiac sign values.}
}
\description{
Discrete zodiac colors
}
\examples{
library(dplyr)
library(ggplot2)

## Use colors that match with each zodiac sign
## sample_dates comes with the package to demonstrate adding zodiac symbols to dates in 2020, all dates from 2000-01-01 to 2000-12-31 are present
p <-
  sample_dates \%>\%
  ## Sample 100 dates from 2020 to count
  sample_n(100) \%>\%
  mutate(zodiac_sign = zodiac_sign(date)) \%>\%
  ggplot() +
  aes(
    x = zodiac_sign,
    fill = zodiac_sign
  ) +
  geom_bar() +
  scale_fill_zodiac(use_factor_order = TRUE)

## Use the colors in the palette without zodiac signs
p2 <-
  iris \%>\%
  ggplot() +
  aes(
    x = Species,
    fill = Species
  ) +
  geom_bar() +
  scale_fill_zodiac()
}
