% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snake_to.R
\name{snake_to}
\alias{snake_to}
\title{Remove Snake Case}
\usage{
snake_to(object, format = "title", acronyms = NULL,
  names_only = FALSE, ggplot_title = FALSE)
}
\arguments{
\item{object}{data.frame or ggplot. A data.frame will have transformed column names. A ggplot object will have transformed X and Y axes.}

\item{format}{A string. The desired target (default is "title") case with options including:
\itemize{
 \item{`"title"` produces title case}
 \item{`"lower"` produces lower case}
 \item{`"normal"` do not transform the string}
 \item{`"sentence"` produces sentence case}
 \item{`"upper"` produces upper case}
 }}

\item{acronyms}{A Character. Default `NULL`. For when acronyms exist in the column names that need to be capitalized. Pass a character vector for when there is more than one acronym. Upper and/or lower case acronyms in this parameter will be accepted. This will only capitalize the wanted words, words that may contain the acronyms letter will NOT be capitalized.}

\item{names_only}{A Logical. Default `FALSE`. If `TRUE`, `snake_to()` will return a vector of transformed column names.}

\item{ggplot_title}{A Logical. Default `FALSE`. If `TRUE`, `snake_to()` will add a title to the ggplot object that refelcts the clean X and Y axes.}
}
\value{
Returns the data.frame with clean names or a vector of strings (based on the `names_only` parameter.
}
\description{
This is useful for when the user wants to use the columns in a presentitive way. Shiny, Power BI, and RMarkdown are use cases where we may want clean column names for the user to read.
}
\examples{
\dontrun{
# snake to title
.data \%>\%
  snake_to()

# snake to sentence
.data \%>\%
  snake_to(format = "sentence")

# snake to title, keep vector of transformed column names
.data \%>\%
  snake_to(names_only = TRUE)

# snake to title, capitalize acronyms
acronyms <- c("pyr", "cyr")
.data \%>\%
  snake_to(acronyms = acronyms)
}
}
