% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_character.R
\name{summary_character}
\alias{summary_character}
\title{Simple summary of a character type variable}
\usage{
summary_character(.data, .col, round = TRUE)
}
\arguments{
\item{.data}{A tbl.}

\item{.col}{A character column that will be summarized}

\item{round}{A logical. If `TRUE`, the default, it will round all decimals to the hundredths position.}
}
\value{
A summarized tbl.
}
\description{
Simple summary of a character type variable
}
