% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_numeric.R
\name{summary_numeric}
\alias{summary_numeric}
\title{Simple summary of a numeric type variable}
\usage{
summary_numeric(.data, .col, round = TRUE)
}
\arguments{
\item{.data}{A tbl.}

\item{.col}{A numeric column that will be summarized}

\item{round}{A logical. If `TRUE`, the default, it will round all decimals to the hundredths position.}
}
\value{
A summarized tbl
}
\description{
With the given numeric column this function will return the min, Q1, median (Q2), Q3, max, mean, sd, and count of missing data.
}
