% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_kow.R
\name{theme_kow}
\alias{theme_kow}
\title{Kow Theme for ggplot2}
\usage{
theme_kow(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Kow Theme for ggplot2
}
\concept{themes kow}
