% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_exists.R
\name{url_exists}
\alias{url_exists}
\title{Check if a URL is accessible}
\usage{
url_exists(x, ..., non_2xx_return_value = FALSE, quiet = FALSE)
}
\arguments{
\item{x}{a single URL}

\item{...}{other params (\code{timeout()} would be a good one) passed directly to \code{httr::HEAD()} and/or \code{httr::GET()}}

\item{non_2xx_return_value}{what to do if the site exists but the HTTP status code is not in the \verb{2xx} range. Default is to return \code{FALSE}.}

\item{quiet}{if not \code{FALSE}, then every time the \code{non_2xx_return_value} condition arises a warning message will be displayed. Default is \code{FALSE}.}
}
\description{
This is useful for checking if URLs are accessible.
}
\references{
https://stackoverflow.com/questions/52911812/check-if-url-exists-in-r
}
