% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_kow_utils.R
\name{use_kowr_utils}
\alias{use_kowr_utils}
\title{Import Utility Functions}
\usage{
use_kowr_utils(name = "utils_kowr")
}
\arguments{
\item{name}{A character. Name of the file.}
}
\description{
Import Utility Functions
}
