% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_html.R
\name{view_html}
\alias{view_html}
\title{View HTML Page}
\usage{
view_html(p)
}
\arguments{
\item{p}{A character vector of HTML code or an xml document that has been created with \code{xml2::read_html()}.}
}
\description{
Handy function to see what the current HTML code looks like as a webpage. I am not the original maker of this function., this is code adapted from reading rvest sessions. Thank you to Lorenzo Gaborini for posting this on their blog! \url{https://adventuresindata.netlify.com/post/2018-01-17-viewing-rvest-sessions/}
}
