% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_rvest_session.R
\name{view_rvest}
\alias{view_rvest}
\title{View rvest Session}
\usage{
view_rvest(sesssion)
}
\arguments{
\item{sesssion}{A rvest session object created with \code{rvest::html_session()}.}
}
\description{
Handy function to see where the current rvest session is. I am not the original maker of this function. Thank you to Lorenzo Gaborini for posting this on their blog! \url{https://adventuresindata.netlify.com/post/2018-01-17-viewing-rvest-sessions/}
}
