% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zodiac_sign.R
\name{zodiac_color_palette}
\alias{zodiac_color_palette}
\title{Zodiac color palette}
\usage{
zodiac_color_palette()
}
\description{
Color palettes for the zodiac signs.
This palette includes 12 colors.
}
\concept{zodiac}
