% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zodiac_sign.R
\name{zodiac_factor_order}
\alias{zodiac_factor_order}
\title{Turn zodiac character into an ordered factor}
\usage{
zodiac_factor_order(x, factor_order = "roman")
}
\arguments{
\item{x}{a vector of data, usually taking a small number of distinct
    values.}

\item{factor_order}{A character. Sort the factor levels by the following:
\itemize{
\item roman: roman calender; Aries first (Default)
\item modern: modern calender; Capricorn first
\item none: return the data as a character
}}
}
\value{
A character with the requests levels.
}
\description{
Turn zodiac character into an ordered factor
}
\examples{
zodiac_factor_order("Libra")
}
