% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zodiac_sign.R
\name{zodiac_sign}
\alias{zodiac_sign}
\title{Extract zodiac sign from date}
\usage{
zodiac_sign(date, factor_order = "roman")
}
\arguments{
\item{date}{A character/date in "yyyy-mm-dd" format. If value is a character the functuin will run \code{lubridate::ymd()} on the \code{date} value.}

\item{factor_order}{A character. Sort the factor levels by the following:
\itemize{
\item roman: roman calender; Aries first (Default)
\item modern: modern calender; Capricorn first
\item none: return the data as a character
}}
}
\value{
A character value. The zodiac sign that relates to the given date
}
\description{
Extract zodiac sign from date
}
\examples{
zodiac_sign("2021-07-03")
}
