% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kowr-data.R
\docType{data}
\name{zodiac_sign_colors}
\alias{zodiac_sign_colors}
\title{Colors that relate to zodiac signs}
\format{
A named vector of length 12.
}
\source{
\url{https://github.com/KoderKow/kowr/blob/master/data-raw/zodiac.R}
}
\usage{
zodiac_sign_colors
}
\description{
A vector of hex colors as values and zodiac as the name value. These colors were hand picked by \href{https://twitter.com/actualtoilet}{ActualToilet} and \href{https://twitter.com/koderkow}{KoderKow}.
}
\keyword{datasets}
