% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kowr-data.R
\docType{data}
\name{zodiac_signs_modern_order}
\alias{zodiac_signs_modern_order}
\title{Zodiac signs modern order}
\format{
A vector of length 12.
}
\source{
\url{https://github.com/KoderKow/kowr/blob/master/data-raw/zodiac.R}
}
\usage{
zodiac_signs_modern_order
}
\description{
Zodiac signs based on the modern calendar year, Jan-Dec.
}
\keyword{datasets}
