% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snake_to.R
\name{snake_to}
\alias{snake_to}
\title{Remove Snake Case}
\usage{
snake_to(dat, format = "title", names_only = FALSE)
}
\arguments{
\item{dat}{A data.frame. The input data.frame.}

\item{format}{A string. The desired target (default is "title") case with options including:
\itemize{
 \item{`"title"` produces title case}
 \item{`"lower"` produces lower case}
 \item{`"normal"` do not transform the string}
 \item{`"sentence"` produces sentence case}
 \item{`"upper"` produces upper case}
 }}

\item{names_only}{A Logical. Default `FALSE`. If `TRUE`, `snake_to()` will return a vector of transformed column names.}
}
\value{
Returns the data.frame with clean names or a vector of strings (based on the `names_only` parameter.
}
\description{
This is useful for when the user wants to use the columns in a presentitive way. Shiny, Power BI, and RMarkdown are use cases where we may want clean column names for the user to read.
}
\examples{
\dontrun{
# snake to title
dat \%>\%
  snake_to()

# snake to sentence
dat \%>\%
  snake_to(format = "sentence")

# snake to title, keep vector of transformed column names
dat \%>\%
  snake_to(names_only = TRUE)
}
}
